Modificación en los comandos para que sea necesario el rol "MOD" para poder ejecutar estos

+++ comandos/canales/quitarCanales.js
@@ -1,6 +1,13 @@
 module.exports = {quitarCanales};
 
 function quitarCanales(mensaje){
+    let mod = mensaje.guild.roles.cache.find(role => role.name === "MOD");
+    if (!mensaje.member.roles.cache.has(mod.id)) {
+        mensaje.reply({
+            content: "No tienes los roles necesarios para ejectuar este comando"
+        })
+        return
+    }
     try {
         let canal = mensaje.mentions.channels.first();
         canal.delete();
